+++ api/compose.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/rand"
+	"path"
 	"strconv"
 	"strings"
 	"time"
@@ -20,7 +21,8 @@ func mapComposeServiceToMesosTask(service cfg.Service, data cfg.Compose, vars ma
 	var cmd mesosutil.Command
 
 	// if task is set then its not a new task and we have to save old needed parameter
-	newTaskID, _ := util.GenUUID()
+	uuid, _ := util.GenUUID()
+	newTaskID := vars["project"] + "_" + name + "." + uuid
 	if task.TaskID != "" {
 		newTaskID = task.TaskID
 		cmd.State = task.State
@@ -30,6 +32,7 @@ func mapComposeServiceToMesosTask(service cfg.Service, data cfg.Compose, vars ma
 	cmd.TaskName = config.PrefixTaskName + ":" + vars["project"] + ":" + name
 	cmd.CPU = getCPU(service)
 	cmd.Memory = getMemory(service)
+	cmd.Disk = getDisk(service)
 	cmd.ContainerType = getLabelValueByKey("biz.aventer.mesos_compose.container_type", service)
 	cmd.ContainerImage = service.Image
 	cmd.NetworkMode = service.NetworkMode
@@ -44,7 +47,8 @@ func mapComposeServiceToMesosTask(service cfg.Service, data cfg.Compose, vars ma
 	cmd.Hostname = getHostname(service)
 	cmd.Command = getCommand(service)
 	cmd.Labels = getLabels(service)
-	cmd.DockerPortMappings = getDockerPorts(service)
+	cmd.Executor = getExecutor(service)
+	cmd.DockerPortMappings = getDockerPorts(service, cmd.Agent)
 	cmd.Environment.Variables = getEnvironment(service)
 	cmd.Volumes = getVolumes(service, data)
 
@@ -70,7 +74,6 @@ func mapComposeServiceToMesosTask(service cfg.Service, data cfg.Compose, vars ma
 	if err != nil {
 		logrus.Error("Cloud not store Mesos Task in Redis: ", err)
 	}
-
 }
 
 // Get the CPU value from the compose file, or the default one if it's unset
@@ -89,6 +92,12 @@ func getMemory(service cfg.Service) float64 {
 		return mem
 	}
 	return config.Memory
+}
+
+// Get the Disk value from the compose file, or the default one if it's unset
+func getDisk(service cfg.Service) float64 {
+	// Currently, onyl default value is supported
+	return config.Disk
 }
 
 // Get the Hostname value from the compose file, or generate one if it's unset
@@ -119,10 +128,42 @@ func getCommand(service cfg.Service) string {
 }
 
 // Get random hostportnumber
-func getRandomHostPort(service cfg.Service) int {
+func getRandomHostPort(agent string) int {
 	rand.Seed(time.Now().UnixNano())
+	// #nosec G404
 	v := rand.Intn(framework.PortRangeTo-framework.PortRangeFrom) + framework.PortRangeFrom
+	port := uint32(v)
+	if portInUse(port, agent) {
+		v = getRandomHostPort(agent)
+	}
 	return v
+}
+
+// Check if the port is already in use
+func portInUse(port uint32, agent string) bool {
+	// get all running services
+	logrus.Debug("Check if port is in use: ", port, agent)
+	keys := GetAllRedisKeys(framework.FrameworkName + ":*")
+	for keys.Next(config.RedisCTX) {
+		// get the details of the current running service
+		key := GetRedisKey(keys.Val())
+		var task mesosutil.Command
+		json.Unmarshal([]byte(key), &task)
+
+		// check if its the given agent
+		if task.Agent == agent {
+			// check if the given port is already in use
+			ports := task.Discovery.GetPorts()
+			if ports == nil {
+				for _, hostport := range ports.GetPorts() {
+					if hostport.Number == port {
+						return true
+					}
+				}
+			}
+		}
+	}
+	return false
 }
 
 // Get the labels of the compose file
@@ -149,25 +190,37 @@ func getLabelValueByKey(label string, service cfg.Service) string {
 }
 
 // Get the ports of the compose file
-func getDockerPorts(service cfg.Service) []mesosproto.ContainerInfo_DockerInfo_PortMapping {
+func getDockerPorts(service cfg.Service, agent string) []mesosproto.ContainerInfo_DockerInfo_PortMapping {
 	var ports []mesosproto.ContainerInfo_DockerInfo_PortMapping
-	hostport := uint32(getRandomHostPort(service))
+	hostport := uint32(getRandomHostPort(agent))
 	for i, c := range service.Ports {
 		var tmp mesosproto.ContainerInfo_DockerInfo_PortMapping
-		p := strings.Split(c, ":")
-		ps, _ := strconv.Atoi(p[1])
-		tmp.ContainerPort = uint32(ps)
-		tmp.HostPort = uint32(hostport + uint32(i))
-		tmp.Protocol = func() *string { x := "tcp"; return &x }()
+		var port int
+		// "<hostport>:<containerport>"
+		// "<hostip>:<hostport>:<containerport>"
+		count := strings.Count(c, ":")
+
+		var p []string
+		var proto []string
+		if count > 0 {
+			p = strings.Split(c, ":")
+			port, _ = strconv.Atoi(p[count])
+			proto = strings.Split(p[count], "/")
+		} else {
+			port, _ = strconv.Atoi(c)
+			proto = strings.Split(c, "/")
+		}
 
 		// check if this is a udp protocol
-		proto := strings.Split(p[1], "/")
+		tmp.Protocol = func() *string { x := "tcp"; return &x }()
 		if len(proto) > 1 {
 			if strings.ToLower(proto[1]) == "udp" {
 				tmp.Protocol = func() *string { x := "udp"; return &x }()
 			}
 		}
 
+		tmp.ContainerPort = uint32(port)
+		tmp.HostPort = hostport + uint32(i)
 		ports = append(ports, tmp)
 	}
 	return ports
@@ -175,26 +228,32 @@ func getDockerPorts(service cfg.Service) []mesosproto.ContainerInfo_DockerInfo_P
 
 // Get the discoveryinfo ports of the compose file
 func getDiscoveryInfoPorts(service cfg.Service, cmd mesosutil.Command) []mesosproto.Port {
-	var port []mesosproto.Port
+	var disport []mesosproto.Port
 	for _, c := range service.Ports {
-		var tmp mesosproto.Port
-		p := strings.Split(c, ":")
-		if len(p) != 2 {
-			continue
+		var tmpport mesosproto.Port
+		var port int
+		// "<hostport>:<containerport>"
+		// "<hostip>:<hostport>:<containerport>"
+		count := strings.Count(c, ":")
+		var p []string
+		if count > 0 {
+			p = strings.Split(c, ":")
+			port, _ = strconv.Atoi(p[count])
+		} else {
+			port, _ = strconv.Atoi(c)
 		}
-		// thats the containerport
-		ps, _ := strconv.Atoi(p[1])
+
 		// create the name of the port
-		name := cmd.TaskName + ":" + p[1]
+		name := cmd.TaskName + ":" + p[count]
 
 		// get the random hostport
-		tmp.Number, tmp.Protocol = getHostPortByContainerPort(ps, cmd)
-		tmp.Name = func() *string { x := name; return &x }()
+		tmpport.Number, tmpport.Protocol = getHostPortByContainerPort(port, cmd)
+		tmpport.Name = func() *string { x := name; return &x }()
 
-		port = append(port, tmp)
+		disport = append(disport, tmpport)
 	}
 
-	return port
+	return disport
 }
 
 // get the random hostport and protcol of the container port
@@ -257,4 +316,43 @@ func getVolumes(service cfg.Service, data cfg.Compose) []mesosproto.Volume {
 		volume = append(volume, tmp)
 	}
 	return volume
+}
+
+// Get custome executer
+func getExecutor(service cfg.Service) mesosproto.ExecutorInfo {
+	var executorInfo mesosproto.ExecutorInfo
+	command := getLabelValueByKey("biz.aventer.mesos_compose.executor", service)
+	uri := getLabelValueByKey("biz.aventer.mesos_compose.executor_uri", service)
+
+	if command != "" {
+		executorID, _ := util.GenUUID()
+		executorInfo = mesosproto.ExecutorInfo{
+			Name: func() *string { x := path.Base(command); return &x }(),
+			Type: mesosproto.ExecutorInfo_CUSTOM,
+			ExecutorID: &mesosproto.ExecutorID{
+				Value: executorID,
+			},
+			FrameworkID: framework.FrameworkInfo.ID,
+			Command: &mesosproto.CommandInfo{
+				Value: func() *string { x := command; return &x }(),
+			},
+			Container: &mesosproto.ContainerInfo{
+				Type: mesosproto.ContainerInfo_MESOS.Enum(),
+			},
+		}
+
+		if uri != "" {
+			executorInfo.Command.URIs = []mesosproto.CommandInfo_URI{
+				{
+					Value:      uri,
+					Extract:    func() *bool { x := false; return &x }(),
+					Executable: func() *bool { x := true; return &x }(),
+					Cache:      func() *bool { x := false; return &x }(),
+					OutputFile: func() *string { x := path.Base(command); return &x }(),
+				},
+			}
+		}
+
+	}
+	return executorInfo
 }