+++ app.go
@@ -2,7 +2,10 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
+	"encoding/base64"
 	"encoding/json"
+	"flag"
 	"fmt"
 	"net/http"
 
@@ -14,24 +17,49 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// MinVersion is the version number of this program
-var MinVersion string
+// BuildVersion of m3s
+var BuildVersion string
+
+// GitVersion is the revision and commit number
+var GitVersion string
 
 // init the redis cache
 func initCache() {
-	client := goredis.NewClient(&goredis.Options{
-		Addr: config.RedisServer,
-		DB:   1,
-	})
+	var redisOptions goredis.Options
+	redisOptions.Addr = config.RedisServer
+	redisOptions.DB = config.RedisDB
+	if config.RedisPassword != "" {
+		redisOptions.Password = config.RedisPassword
+	}
+	client := goredis.NewClient(&redisOptions)
+
 	config.RedisCTX = context.Background()
 	pong, err := client.Ping(config.RedisCTX).Result()
 	logrus.Debug("Redis Health: ", pong, err)
 	config.RedisClient = client
 }
 
+// convert Base64 Encodes PEM Certificate to tls object
+func decodeBase64Cert(pemCert string) []byte {
+	sslPem, err := base64.URLEncoding.DecodeString(pemCert)
+	if err != nil {
+		logrus.Fatal("Error decoding SSL PEM from Base64: ", err.Error())
+	}
+	return sslPem
+}
+
 func main() {
+	// Prints out current version
+	var version bool
+	flag.BoolVar(&version, "v", false, "Prints current version")
+	flag.Parse()
+	if version {
+		fmt.Print(GitVersion)
+		return
+	}
+
 	util.SetLogging(config.LogLevel, config.EnableSyslog, config.AppName)
-	logrus.Println(config.AppName + " build " + MinVersion)
+	logrus.Println(config.AppName + " build " + BuildVersion + " git " + GitVersion)
 
 	listen := fmt.Sprintf(":%s", framework.FrameworkPort)
 
@@ -42,6 +70,7 @@ func main() {
 	framework.CommandChan = make(chan mesosutil.Command, 100)
 	config.Hostname = framework.FrameworkHostname
 	config.Listen = listen
+	config.Suppress = false
 
 	framework.State = map[string]mesosutil.State{}
 
@@ -56,30 +85,50 @@ func main() {
 	//		{Type: mesosproto.FrameworkInfo_Capability_RESERVATION_REFINEMENT},
 	//	}
 
-	// The Hostname should ever be set after reading the state file.
-	framework.FrameworkInfo.Hostname = &framework.FrameworkHostname
-
 	initCache()
+
 	mesosutil.SetConfig(&framework)
 	api.SetConfig(&config, &framework)
 	mesos.SetConfig(&config, &framework)
 
 	// load framework state from database if they exist
-	key := api.GetRedisKey("framework")
+	key := api.GetRedisKey(framework.FrameworkName + ":framework")
 	if key != "" {
 		json.Unmarshal([]byte(key), &framework)
+
+		// Save current config
+		api.SaveConfig()
 	}
 
-	// load framework config from DB
-	key = api.GetRedisKey("framework_config")
-	if key != "" {
-		json.Unmarshal([]byte(key), &config)
+	// The Hostname should ever be set after reading the state file.
+	framework.FrameworkInfo.Hostname = &framework.FrameworkHostname
+
+	server := &http.Server{
+		Addr:    listen,
+		Handler: api.Commands(),
+		TLSConfig: &tls.Config{
+			ClientAuth: tls.RequestClientCert,
+			MinVersion: tls.VersionTLS12,
+		},
 	}
 
-	http.Handle("/", api.Commands())
+	if config.SSLCrt != "" && config.SSLKey != "" {
+		logrus.Debug("Enable TLS")
+		crt := decodeBase64Cert(config.SSLCrt)
+		key := decodeBase64Cert(config.SSLKey)
+		certs, err := tls.X509KeyPair(crt, key)
+		if err != nil {
+			logrus.Fatal("TLS Server Error: ", err.Error())
+		}
+		server.TLSConfig.Certificates = []tls.Certificate{certs}
+	}
 
 	go func() {
-		http.ListenAndServe(listen, nil)
+		if config.SSLCrt != "" && config.SSLKey != "" {
+			server.ListenAndServeTLS("", "")
+		} else {
+			server.ListenAndServe()
+		}
 	}()
 	logrus.Fatal(mesos.Subscribe())
 }