+++ mesos/mesos.go
@@ -6,7 +6,6 @@ import (
 	"crypto/tls"
 	"encoding/json"
 	"net/http"
-	"strconv"
 	"strings"
 
 	api "github.com/AVENTER-UG/mesos-compose/api"
@@ -49,7 +48,8 @@ func Subscribe() error {
 	logrus.Debug(body)
 	client := &http.Client{}
 	client.Transport = &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+		// #nosec G402
+		TLSClientConfig: &tls.Config{InsecureSkipVerify: config.SkipSSL},
 	}
 
 	protocol := "https"
@@ -65,22 +65,20 @@ func Subscribe() error {
 	if err != nil {
 		logrus.Fatal(err)
 	}
+	defer res.Body.Close()
 
 	reader := bufio.NewReader(res.Body)
 
 	line, _ := reader.ReadString('\n')
-	bytesCount, _ := strconv.Atoi(strings.Trim(line, "\n"))
+	line = strings.TrimSuffix(line, "\n")
 
 	for {
 		// Read line from Mesos
 		line, _ = reader.ReadString('\n')
-		line = strings.Trim(line, "\n")
+		line = strings.TrimSuffix(line, "\n")
 		// Read important data
-		data := line[:bytesCount]
-		// Rest data will be bytes of next message
-		bytesCount, _ = strconv.Atoi((line[bytesCount:]))
 		var event mesosproto.Event // Event as ProtoBuf
-		err := jsonpb.UnmarshalString(data, &event)
+		err := jsonpb.UnmarshalString(line, &event)
 		if err != nil {
 			logrus.Error(err)
 		}
@@ -96,17 +94,15 @@ func Subscribe() error {
 
 			// store framework configuration
 			d, _ := json.Marshal(&framework)
-			err := config.RedisClient.Set(config.RedisCTX, "framework", d, 0).Err()
+			err = config.RedisClient.Set(config.RedisCTX, framework.FrameworkName+":framework", d, 0).Err()
 			if err != nil {
 				logrus.Error("Framework save config and state into redis Error: ", err)
 			}
 			Reconcile()
+			api.SaveConfig()
 		case mesosproto.Event_UPDATE:
 			logrus.Debug("Update", HandleUpdate(&event))
-			err := api.SaveConfig()
-			if err != nil {
-				api.ErrorMessage(1, "Event_UPDATE", "Could not save config data")
-			}
+			api.SaveConfig()
 		case mesosproto.Event_HEARTBEAT:
 			Heartbeat()
 		case mesosproto.Event_OFFERS: