+++ mesos/prepare.go
@@ -10,11 +10,13 @@ import (
 )
 
 func defaultResources(cmd mesosutil.Command) []mesosproto.Resource {
+	PORT := "ports"
 	CPU := "cpus"
 	MEM := "mem"
+	DISK := "disk"
 	cpu := cmd.CPU
 	mem := cmd.Memory
-	PORT := "ports"
+	disk := cmd.Disk
 
 	res := []mesosproto.Resource{
 		{
@@ -27,6 +29,11 @@ func defaultResources(cmd mesosutil.Command) []mesosproto.Resource {
 			Type:   mesosproto.SCALAR.Enum(),
 			Scalar: &mesosproto.Value_Scalar{Value: mem},
 		},
+		{
+			Name:   DISK,
+			Type:   mesosproto.SCALAR.Enum(),
+			Scalar: &mesosproto.Value_Scalar{Value: disk},
+		},
 	}
 
 	var portBegin, portEnd uint64
@@ -35,40 +42,30 @@ func defaultResources(cmd mesosutil.Command) []mesosproto.Resource {
 		portBegin = uint64(cmd.DockerPortMappings[0].HostPort)
 		portEnd = portBegin + uint64(len(cmd.DockerPortMappings)) - 1
 
-		res = []mesosproto.Resource{
-			{
-				Name:   CPU,
-				Type:   mesosproto.SCALAR.Enum(),
-				Scalar: &mesosproto.Value_Scalar{Value: cpu},
-			},
-			{
-				Name:   MEM,
-				Type:   mesosproto.SCALAR.Enum(),
-				Scalar: &mesosproto.Value_Scalar{Value: mem},
-			},
-			{
-				Name: PORT,
-				Type: mesosproto.RANGES.Enum(),
-				Ranges: &mesosproto.Value_Ranges{
-					Range: []mesosproto.Value_Range{
-						{
-							Begin: portBegin,
-							End:   portEnd,
-						},
+		port := mesosproto.Resource{
+			Name: PORT,
+			Type: mesosproto.RANGES.Enum(),
+			Ranges: &mesosproto.Value_Ranges{
+				Range: []mesosproto.Value_Range{
+					{
+						Begin: portBegin,
+						End:   portEnd,
 					},
 				},
 			},
 		}
+		res = append(res, port)
 	}
 
 	return res
 }
 
+// PrepareTaskInfoExecuteContainer will create the TaskInfo Protobuf for Mesos
 func PrepareTaskInfoExecuteContainer(agent mesosproto.AgentID, cmd mesosutil.Command) ([]mesosproto.TaskInfo, error) {
+	contype := mesosproto.ContainerInfo_DOCKER.Enum()
+
 	d, _ := json.Marshal(&cmd)
 	logrus.Debug("HandleOffers cmd: ", util.PrettyJSON(d))
-
-	contype := mesosproto.ContainerInfo_DOCKER.Enum()
 
 	// Set Container Network Mode
 	networkMode := mesosproto.ContainerInfo_DockerInfo_BRIDGE.Enum()
@@ -110,20 +107,24 @@ func PrepareTaskInfoExecuteContainer(agent mesosproto.AgentID, cmd mesosutil.Com
 		}
 	}
 
-	msg.Container = &mesosproto.ContainerInfo{}
-	msg.Container.Type = contype
-	msg.Container.Volumes = cmd.Volumes
-	msg.Container.Docker = &mesosproto.ContainerInfo_DockerInfo{
-		Image:        cmd.ContainerImage,
-		Network:      networkMode,
-		PortMappings: cmd.DockerPortMappings,
-		Privileged:   &cmd.Privileged,
-		Parameters:   cmd.DockerParameter,
-	}
-	msg.Container.NetworkInfos = cmd.NetworkInfo
+	// ExecutorInfo or CommandInfo/Container, both is not supportet
+	if getLabelValue("biz.aventer.mesos_compose.executor", cmd) == "" {
+		msg.Container = &mesosproto.ContainerInfo{}
+		msg.Container.Type = contype
+		msg.Container.Volumes = cmd.Volumes
+		msg.Container.Docker = &mesosproto.ContainerInfo_DockerInfo{
+			Image:          cmd.ContainerImage,
+			Network:        networkMode,
+			PortMappings:   cmd.DockerPortMappings,
+			Privileged:     &cmd.Privileged,
+			Parameters:     cmd.DockerParameter,
+			ForcePullImage: func() *bool { x := true; return &x }(),
+		}
+		msg.Container.NetworkInfos = cmd.NetworkInfo
 
-	if cmd.Hostname != "" {
-		msg.Container.Hostname = &cmd.Hostname
+		if cmd.Hostname != "" {
+			msg.Container.Hostname = &cmd.Hostname
+		}
 	}
 
 	if cmd.Discovery != (mesosproto.DiscoveryInfo{}) {
@@ -135,9 +136,6 @@ func PrepareTaskInfoExecuteContainer(agent mesosproto.AgentID, cmd mesosutil.Com
 			Labels: cmd.Labels,
 		}
 	}
-
-	d, _ = json.Marshal(&msg)
-	logrus.Debug("HandleOffers msg: ", util.PrettyJSON(d))
 
 	return []mesosproto.TaskInfo{msg}, nil
 }