+++ types/types.go
@@ -21,9 +21,17 @@ type Config struct {
 	PrefixTaskName string
 	CPU            float64
 	Memory         float64
+	Disk           float64
 	RedisServer    string
 	RedisClient    *goredis.Client
 	RedisCTX       context.Context
+	RedisPassword  string
+	RedisDB        int
+	SkipSSL        bool
+	SSLKey         string
+	SSLCrt         string
+	Suppress       bool
+	Agents         MesosAgent
 }
 
 // UserCredentials - The Username and Password to authenticate against this framework
@@ -32,7 +40,7 @@ type UserCredentials struct {
 	Password string
 }
 
-// Yaml2Go
+// Compose - The main structure of the supported docker-compose syntax
 type Compose struct {
 	Version  string              `yaml:"version"`
 	Services map[string]Service  `yaml:"services"`
@@ -40,7 +48,7 @@ type Compose struct {
 	Volumes  map[string]Volumes  `yaml:"volumes"`
 }
 
-// Web
+// Service - The docker-compose service parameters
 type Service struct {
 	Network     []string               `yaml:"network"`
 	Build       string                 `yaml:"build"`
@@ -52,12 +60,13 @@ type Service struct {
 	Image       string                 `yaml:"image"`
 	Labels      map[string]interface{} `yaml:"labels"`
 	NetworkMode string                 `yaml:"network_mode"`
-	Privileged  bool                   `yaml:"priviliged"`
+	Privileged  bool                   `yaml:"privileged"`
 	Command     []string               `yaml:"command"`
 	Deploy      Deploy                 `yaml:"deploy"`
 	Hostname    string                 `yaml:"hostname"`
 }
 
+// Deploy - The mesos resources to deploy a task
 type Deploy struct {
 	Resources struct {
 		Limits struct {
@@ -67,12 +76,14 @@ type Deploy struct {
 	} `yaml:"resources"`
 }
 
+// Networks - The docker-compose network syntax
 type Networks struct {
 	External bool   `yaml:"external"`
 	Name     string `yaml:"name"`
 	Driver   string `yaml:"driver"`
 }
 
+// Volumes - The docker-compose volumes syntax
 type Volumes struct {
 	Driver string `yaml:"driver"`
 }