+++ changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v0.3.0
+
+- Add Redis Authentication Support
+- Change DB items framework and framework_config to be saved with the
+  frameworkName as prefix.
+- The default prefix of hostname and tasks are the FrameworkName.
+- Optimize framework suppress
+- Add constraint "hostname eq" to accept only offers with the given hostname.
+- Check if the random generated port is already in use at the given agent.
+- Change TaskID format to be more speakable.
+- IMPORTANT!!! Change the API URL's.
+- Add force pull image
+- Add executor label to define custom mesos executor.
+
 ## v0.2.0
 
 - Start with changelog
@@ -8,15 +22,13 @@
 - Add Docker Volume support
 - Add Environment variables
 - Add Restart of Mesos Tasks
-- Add Kill Service 
+- Add Kill Service
 - Add Kill single task of a service
 - Add Framework Suppress if there is nothing to schedule
 - Remove tasks after they are killed
 - Add reconcile after frameware resubscribe
+- Add TLS Server Support (env variable SSL_CRT_BASE64, SSL_KEY_BASE64)
 
-
-## v0.1.0 
+## v0.1.0
 
 - First inn
-
-